added seeds, refactored helpers and routers
